requirements.txt
python-telegram-bot
pypdf
openai
python-dotenv
requests